Fix compatibility with new dependencies
- Fix RoutingMark type from int to uint32
- Fix ech_quic.go: Update CreateTransport signature and QUICConfig field name
- Fix geosite/reader.go: Replace ReadCounter with custom readCounter wrapper
- Fix srs/binary.go: Add varbinReader/varbinWriter wrappers for domain matcher
- Fix fakeip/server.go: Wrap NewTransport function for dns.RegisterTransport

+++ common/dialer/default.go
@@ -45,11 +45,11 @@ func NewDefault(router adapter.Router, options option.DialerOptions) (*DefaultDi
 		listener.Control = control.Append(listener.Control, bindFunc)
 	}
 	if options.RoutingMark != 0 {
-		dialer.Control = control.Append(dialer.Control, control.RoutingMark(options.RoutingMark))
-		listener.Control = control.Append(listener.Control, control.RoutingMark(options.RoutingMark))
+		dialer.Control = control.Append(dialer.Control, control.RoutingMark(uint32(options.RoutingMark)))
+		listener.Control = control.Append(listener.Control, control.RoutingMark(uint32(options.RoutingMark)))
 	} else if router.DefaultMark() != 0 {
-		dialer.Control = control.Append(dialer.Control, control.RoutingMark(router.DefaultMark()))
-		listener.Control = control.Append(listener.Control, control.RoutingMark(router.DefaultMark()))
+		dialer.Control = control.Append(dialer.Control, control.RoutingMark(uint32(router.DefaultMark())))
+		listener.Control = control.Append(listener.Control, control.RoutingMark(uint32(router.DefaultMark())))
 	}
 	if options.ReuseAddr {
 		listener.Control = control.Append(listener.Control, control.ReuseAddr())
